+++ src/steps/resource-manager/security/client.ts
@@ -1,5 +1,6 @@
 import { SecurityCenter } from '@azure/arm-security';
 import {
+  Pricing,
   SecurityAssessment,
   SecurityContact,
 } from '@azure/arm-security/esm/models';
@@ -29,6 +30,30 @@ export class SecurityClient extends Client {
         },
       },
       resourceDescription: 'security.assessments',
+      callback,
+    });
+  }
+
+  public async iteratePricings(
+    callback: (p: Pricing) => void | Promise<void>,
+  ): Promise<void> {
+    const serviceClient = await this.getAuthenticatedServiceClient(
+      SecurityCenter,
+    );
+
+    return iterateAllResources({
+      logger: this.logger,
+      serviceClient,
+      resourceEndpoint: {
+        list: async () => {
+          const response = await serviceClient.pricings.list();
+          const pricings = response.value;
+          return Object.assign(pricings, {
+            _response: response._response,
+          });
+        },
+      },
+      resourceDescription: 'security.pricings',
       callback,
     });
   }