+++ src/steps/resource-manager/security/index.test.ts
@@ -1,11 +1,22 @@
-import { fetchAssessments, fetchSecurityCenterContacts } from '.';
+import {
+  fetchAssessments,
+  fetchSecurityCenterContacts,
+  fetchSecurityCenterPricingConfigurations,
+} from '.';
 import { Recording } from '@jupiterone/integration-sdk-testing';
 import { IntegrationConfig } from '../../../types';
-import { setupAzureRecording } from '../../../../test/helpers/recording';
+import {
+  getMatchRequestsBy,
+  setupAzureRecording,
+} from '../../../../test/helpers/recording';
 import { entities as subscriptionEntities } from '../subscriptions/constants';
 import { ACCOUNT_ENTITY_TYPE } from '../../active-directory';
-import { SecurityEntities } from './constants';
+import { SecurityEntities, SecurityRelationships } from './constants';
 import { createMockAzureStepExecutionContext } from '../../../../test/createMockAzureStepExecutionContext';
+import { configFromEnv } from '../../../../test/integrationInstanceConfig';
+import { getMockAccountEntity } from '../../../../test/helpers/getMockAccountEntity';
+import { v4 as uuid } from 'uuid';
+
 let recording: Recording;
 
 afterEach(async () => {
@@ -101,4 +112,87 @@ test('step - security center contacts', async () => {
       displayName: 'HAS',
     }),
   );
+});
+
+describe('rm-security-center-pricing-configs', () => {
+  function getSetupEntities(config: IntegrationConfig) {
+    const accountEntity = getMockAccountEntity(config);
+
+    const subscriptionId = `/subscriptions/${config.subscriptionId}`;
+    const subscriptionEntity = {
+      _key: subscriptionId,
+      _type: 'subscription-type',
+      _class: 'subscription-class',
+    };
+
+    return { accountEntity, subscriptionEntity };
+  }
+
+  function getConfigForTest(config: IntegrationConfig): IntegrationConfig {
+    // The Azure Subscription Client has some validation before sending requests to
+    // this endpoint that requires "subscriptionId" to be a UUID.
+    //
+    // When running tests in CI, we set the value of `config.subscriptionId` to "subscriptionId",
+    // causing the Azure SDK to throw the following:
+    //
+    // "subscriptionId" with value "subscriptionId" should satisfy the constraint
+    // "Pattern": /^[0-9A-Fa-f]{8}-([0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$/.
+
+    const subscriptionIdValidationRegex = /^[0-9A-Fa-f]{8}-([0-9A-Fa-f]{4}-){3}[0-9A-Fa-f]{12}$/;
+    return {
+      ...config,
+      subscriptionId: subscriptionIdValidationRegex.test(
+        config.subscriptionId || '',
+      )
+        ? config.subscriptionId
+        : uuid(),
+    };
+  }
+
+  test('success', async () => {
+    const config = getConfigForTest(configFromEnv);
+
+    recording = setupAzureRecording({
+      directory: __dirname,
+      name: 'rm-security-center-pricing-configs',
+      options: {
+        matchRequestsBy: getMatchRequestsBy({ config }),
+      },
+    });
+
+    const { accountEntity, subscriptionEntity } = getSetupEntities(config);
+
+    const context = createMockAzureStepExecutionContext({
+      instanceConfig: config,
+      entities: [subscriptionEntity],
+      setData: {
+        [ACCOUNT_ENTITY_TYPE]: accountEntity,
+      },
+    });
+
+    await fetchSecurityCenterPricingConfigurations(context);
+
+    const pricingEntities = context.jobState.collectedEntities;
+
+    expect(pricingEntities.length).toBeGreaterThan(0);
+    expect(pricingEntities).toMatchGraphObjectSchema({
+      _class: SecurityEntities.SUBSCRIPTION_PRICING._class,
+    });
+
+    const subscriptionPricingRelationships =
+      context.jobState.collectedRelationships;
+
+    expect(subscriptionPricingRelationships.length).toBe(
+      pricingEntities.length,
+    );
+    expect(subscriptionPricingRelationships).toMatchDirectRelationshipSchema({
+      schema: {
+        properties: {
+          _type: {
+            const: SecurityRelationships.SUBSCRIPTION_HAS_PRICING_CONFIG._type,
+          },
+        },
+      },
+    });
+  });
 });